refactor circle drawing methods a bit
