+++ src/models/Warp.js
@@ -77,7 +77,7 @@ export default class Warp extends Effect {
 
   getVertices(state) {
     const width = state.shape.startingWidth
-    return circle(width/2, 0)
+    return circle(width/2)
   }
 
   applyEffect(effect, layer, vertices) {
